делит код на модули по-другому
